chore: implemented TransactionHeader component
